refactor: move mapping to validation page
